feat: enable more than 255 users at the same time

+++ Server/Network/Handlers/EditorHandler.cs
@@ -58,7 +58,7 @@ internal static class EditorHandler
         ClassRepository.WriteAll();
 
         // Envia os novos dados para todos jogadores conectados
-        for (byte i = 0; i < Account.List.Count; i++)
+        for (int i = 0; i < Account.List.Count; i++)
             if (Account.List[i] != account)
                 ClassSender.Classes(Account.List[i]);
     }
@@ -83,7 +83,7 @@ internal static class EditorHandler
             tempMap.SpawnItems();
 
             // Envia o mapa para todos os jogadores que estão nele
-            for (byte n = 0; n < Account.List.Count; n++)
+            for (int n = 0; n < Account.List.Count; n++)
                 if (Account.List[n] != account)
                     if (Account.List[n].Character.Map == tempMap || Account.List[n].InEditor)
                         MapSender.Map(Account.List[n], tempMap.Data);
@@ -104,7 +104,7 @@ internal static class EditorHandler
         NpcRepository.WriteAll();
 
         // Envia os novos dados para todos jogadores conectados
-        for (byte i = 0; i < Account.List.Count; i++)
+        for (int i = 0; i < Account.List.Count; i++)
             if (Account.List[i] != account)
                 NpcSender.Npcs(Account.List[i]);
     }
@@ -123,7 +123,7 @@ internal static class EditorHandler
         ItemRepository.WriteAll();
 
         // Envia os novos dados para todos jogadores conectados
-        for (byte i = 0; i < Account.List.Count; i++)
+        for (int i = 0; i < Account.List.Count; i++)
             if (Account.List[i] != account)
                 ItemSender.Items(Account.List[i]);
     }
@@ -142,7 +142,7 @@ internal static class EditorHandler
         ShopRepository.WriteAll();
 
         // Envia os novos dados para todos jogadores conectados
-        for (byte i = 0; i < Account.List.Count; i++)
+        for (int i = 0; i < Account.List.Count; i++)
             if (Account.List[i] != account)
                 ShopSender.Shops(Account.List[i]);
     }
