+++ Server/Network/Send.cs
@@ -20,7 +20,7 @@ internal static class Send
     public static void ToAll(NetDataWriter data)
     {
         // Envia os dados para todos conectados
-        for (byte i = 0; i < Account.List.Count; i++)
+        for (int i = 0; i < Account.List.Count; i++)
             if (Account.List[i].IsPlaying)
                 ToPlayer(Account.List[i].Character, data);
     }
@@ -28,7 +28,7 @@ internal static class Send
     public static void ToAllBut(Player player, NetDataWriter data)
     {
         // Envia os dados para todos conectados, com excessão do índice
-        for (byte i = 0; i < Account.List.Count; i++)
+        for (int i = 0; i < Account.List.Count; i++)
             if (Account.List[i].IsPlaying)
                 if (player != Account.List[i].Character)
                     ToPlayer(Account.List[i].Character, data);
@@ -37,7 +37,7 @@ internal static class Send
     public static void ToMap(TempMap map, NetDataWriter data)
     {
         // Envia os dados para todos conectados, com excessão do índice
-        for (byte i = 0; i < Account.List.Count; i++)
+        for (int i = 0; i < Account.List.Count; i++)
             if (Account.List[i].IsPlaying)
                 if (Account.List[i].Character.Map == map)
                     ToPlayer(Account.List[i].Character, data);
@@ -46,7 +46,7 @@ internal static class Send
     public static void ToMapBut(TempMap map, Player player, NetDataWriter data)
     {
         // Envia os dados para todos conectados, com excessão do índice
-        for (byte i = 0; i < Account.List.Count; i++)
+        for (int i = 0; i < Account.List.Count; i++)
             if (Account.List[i].IsPlaying)
                 if (Account.List[i].Character.Map == map)
                     if (player != Account.List[i].Character)
