+++ Server/Program.cs
@@ -67,7 +67,7 @@ internal static class Program
     private static void PerformShutdown()
     {
         // Salva os dados de todos os jogadores
-        for (byte i = 0; i < Account.List.Count; i++)
+        for (int i = 0; i < Account.List.Count; i++)
             if (Account.List[i].IsPlaying)
                 CharacterRepository.Write(Account.List[i]);
